Refactor a widget for scrolling through a list. Use to highlight one movement at a time.
